// web-application/types-openapi/api/models/ListRoleRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListRoleRequest
 */
export interface ListRoleRequest {
    /**
     * 
     * @type {string}
     * @memberof ListRoleRequest
     */
    roleName?: string | null;
}

/**
 * Check if a given object implements the ListRoleRequest interface.
 */
export function instanceOfListRoleRequest(value: object): value is ListRoleRequest {
    return true;
}

export function ListRoleRequestFromJSON(json: any): ListRoleRequest {
    return ListRoleRequestFromJSONTyped(json, false);
}

export function ListRoleRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListRoleRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'roleName': json['roleName'] == null ? undefined : json['roleName'],
    };
}

export function ListRoleRequestToJSON(json: any): ListRoleRequest {
    return ListRoleRequestToJSONTyped(json, false);
}

export function ListRoleRequestToJSONTyped(value?: ListRoleRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roleName': value['roleName'],
    };
}