// web-application/types-openapi/api/models/ListRoleResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListRoleResponse
 */
export interface ListRoleResponse {
    /**
     * 
     * @type {string}
     * @memberof ListRoleResponse
     */
    roleId: string;
    /**
     * 
     * @type {string}
     * @memberof ListRoleResponse
     */
    roleName: string | null;
}

/**
 * Check if a given object implements the ListRoleResponse interface.
 */
export function instanceOfListRoleResponse(value: object): value is ListRoleResponse {
    if (!('roleId' in value) || value['roleId'] === undefined) return false;
    if (!('roleName' in value) || value['roleName'] === undefined) return false;
    return true;
}

export function ListRoleResponseFromJSON(json: any): ListRoleResponse {
    return ListRoleResponseFromJSONTyped(json, false);
}

export function ListRoleResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListRoleResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'roleId': json['roleId'],
        'roleName': json['roleName'],
    };
}

export function ListRoleResponseToJSON(json: any): ListRoleResponse {
    return ListRoleResponseToJSONTyped(json, false);
}

export function ListRoleResponseToJSONTyped(value?: ListRoleResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roleId': value['roleId'],
        'roleName': value['roleName'],
    };
}